pyproject.toml
[project]
name = "analysis-mujoco-simulator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ipython>=8.31.0",
    "mujoco-mjx>=3.2.5",
    "notebook>=7.3.2",
]
